+++ springboot/src/main/java/com/example/tickets/exception/TicketNotFoundException.java
@@ -4,7 +4,7 @@ package com.example.tickets.exception;
 // In Java: extends RuntimeException so Spring maps it to HTTP 404 via @ExceptionHandler
 public class TicketNotFoundException extends RuntimeException {
 
-    public TicketNotFoundException(Long id) {
+    public TicketNotFoundException(Integer id) {
         super("Ticket not found: " + id);
     }
 }
